CLEM Improved relationship components for CLEM
