+++ Models/CLEM/Resources/Labour.cs
@@ -27,7 +27,7 @@ namespace Models.CLEM.Resources
     {
         private List<string> WarningsMultipleEntry = new List<string>();
         private List<string> WarningsNotFound = new List<string>();
-        private LabourAERelationship adultEquivalentRelationship = null;
+        private Relationship adultEquivalentRelationship = null;
 
         /// <summary>
         /// Get the Clock.
@@ -72,7 +72,7 @@ namespace Models.CLEM.Resources
             availabilityList = Apsim.Children(this, typeof(LabourAvailabilityList)).Cast<LabourAvailabilityList>().FirstOrDefault();
 
             // locate AE relationship
-            adultEquivalentRelationship = Apsim.Children(this, typeof(LabourAERelationship)).Cast<LabourAERelationship>().FirstOrDefault();
+            adultEquivalentRelationship = Apsim.Children(this, typeof(Relationship)).Where(a => a.Name.ToUpper().Contains("AE")).Cast<Relationship>().FirstOrDefault();
 
             if (Clock.Today.Day != 1)
             {
