+++ Models/Models.csproj
@@ -163,6 +163,8 @@
     <Compile Include="CLEM\Market.cs" />
     <Compile Include="CLEM\RainfallShuffler.cs" />
     <Compile Include="CLEM\RandomNumberGenerator.cs" />
+    <Compile Include="CLEM\Relationship.cs" />
+    <Compile Include="CLEM\RelationshipTracker.cs" />
     <Compile Include="CLEM\Reporting\ConceptionStatus.cs" />
     <Compile Include="CLEM\Reporting\ConceptionStatusChangedEventArgs.cs" />
     <Compile Include="CLEM\Resources\LabourDietComponent.cs" />
